Add date picker functionality
Why is it necessary? (Bug fix, feature, improvements)
This feature will allow the user select the date for the appointment

How does the change address the issuer?
This feature uses a ScrollView to display the dates along a few VIEW and TEXT components.

CalendarDate produces the days of the weeks using the computers UNIX Timestamp

What side effect does this change have ?
It does not affect any other component.

https://github.com/users/cs16mmp/projects/1#card-33513756