Fix Route Navigation
Why is it necessary? (Bug fix, feature, improvements)
The app realies on this feature to navigate through the different screens

How does the change address the issuer?
It employs Route Navigator - Making Home Screen the principal screen of the app

What side effect does this change have ?
It seems this feature has added a top bar to every screen

https://github.com/users/cs16mmp/projects/1#card-33521337