Add AWS compatibility using Amplify
Why is it necessary? (Bug fix, feature, improvements)
This will be necessary to connect the application to the Amazon Web Server.
There the app will retrieve information about bookings.

How does the change address the issuer?
It does not fully address the issue just yet.
This step only intially connects the app to the server so it can easily access to the services in the future.

What side effect does this change have ?
It should not have any side effect.

https://github.com/users/cs16mmp/projects/1#card-33513777