Add a feature to fill the AWS database with random appointments
Why is it necessary? (Bug fix, feature, improvements)
This will be used to test the AWS database

How does the change address the issuer?
It employs a for loops to generate timestamps

What side effect does this change have ?
Not sure of the side effect just yet

Include a link to the ticket, if any
https://github.com/cs16mmp/dotsEarth/projects/1#card-33714286
https://github.com/cs16mmp/dotsEarth/projects/1#card-33588345