Add REDUX capabilities
Why is it necessary? (Bug fix, feature, improvements)
This could be used in the future to manage the state of the applicaiton and to store data

How does the change address the issuer?
I would be used to address some issues

What side effect does this change have ?
It may increase the complexity of the project