Booking system saves the description of the appointment
Why is it necessary? (Bug fix, feature, improvements)
This feature is link to the booking system. this feature allows the user to save the information of their appointment in a different table

How does the change address the issuer?
It emaploys Dynamo DB connected to AppSync and GraphQL mutations.
Also I have re-engineerd some parts of the code such as the creation of ServerInterface to house all the server code

What side effect does this change have ?
None for now

https://github.com/cs16mmp/dotsEarth/projects/1#card-34029117